ajusta as chamadas das rotas de clientes, testando os payloads

+++ src/services/client.service.ts
@@ -1,5 +1,4 @@
 import { NotFoundError } from "../errors/not-found.error";
-import { ValidationError } from "../errors/validation.error";
 import { Client } from "../models/client.model";
 import { ClientRepository } from "../repositories/client.repository";
 
@@ -30,20 +29,19 @@ export class ClientService {
         client.id = this.newId();
         await this.clientRepository.create(client)
     }
+
     async update(client: Client, clientId: number): Promise<void> {
-        if(this.getTipoCliente(client) === "PF" && client.cnpj !== null ){
-            throw new ValidationError("Cliente do tipo Pessoa Física não pode ter um CNPJ!")
-        }
-        if(this.getTipoCliente(client) === "PJ" && client.cpf !== null ){
-            throw new ValidationError("Cliente do tipo Pessoa Jurídica não pode ter um CPF!")
-        }
         const _client = await this.getById(clientId);
         if (_client) {
+            client.produtosContratados = _client.produtosContratados;
             await this.clientRepository.update(client, clientId);
         }
     }
     async delete(clientId: number): Promise<void> {
+        const _client = await this.getById(clientId);
+        if (_client) {
         await this.clientRepository.delete(clientId);
+        }
     }   
 
     getTipoCliente(client: Client): string {
